requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
pypdf==6.7.5
PyJWT==2.10.1
PyMySQL==1.1.2
reportlab==4.4.10
sqlparse==0.5.3
